requirements.txt
requests
python-dotenv
tqdm
aiohttp